requirements.txt
python-dotenv
pcpartpicker
cohere